Add drive letter to snapshot description
I've added some code sourced from [1] that adds the drive letter of the volume to the description of the snapshot in the format of servername-c:-backup-YYYY-MM-DD. This makes it a bit easier to identify a snapshot for a specific volume, especially if the server has multiple EBS volumes attached, even more so if they are all the same size. 

I've tested this out in my AWS account and it works fine, but not sure what all validation is needed from your side. 

[1] http://docs.aws.amazon.com/AWSEC2/latest/WindowsGuide/ec2-windows-volumes.html